// Project-Web/src/app/openapi-api-models/model/eventDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {TicketDto} from './ticketDto';
import {LocationDto} from "./locationDto";
import {FileHandle} from "./FileHandle";


export interface EventDto {
  id?: number;
  name: string;
  description: string;
  category: string;
  date: string;
  hour?: string;
  organizerId: string;
  availableTickets?: number;
  soldTickets?: number;
  location: LocationDto;
  tickets: Array<TicketDto>;
  image: FileHandle;
  imageString?: string;
}